Support for multiple LDAP servers and connection failover
This commit implements the support for setting multiple LDAP servers,
and to automatically fallback to the next available LDAP server during
the authentication of a user.

This commit also adds the command `LDAP.STATUS` that shows the
connection status for each LDAP server configured in the module.

To implement this feature we started using the async API of the ldap3
library, which required several structural changes to the code.

Fixes: #13, #14, #15

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>