Asynchronous management of LDAP connections
In this commit we made a huge structural change in the way the LDAP
authentication requests are handled.

In this new implementation, the whole workflow of an authentication
request is asynchronous, by making use of the Rust tokio multithreaded
runtime.

The architecture is quite simple. There's a scheduler component that
recieves a task and executes it asynchronously using one of the tokio's
threads. The LDAP connections also make use of async IO as already
supported by the ldap library that we're using.

Besides the asynchronous execution model, we also implemented a
connection pool for each LDAP server present in the configuration.
Currently the number of connections in the pool is static and is
controlled by the `ldap.connection_pool_size` configuration option.

Fixes: #18

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>