Catch and log error when `ldap.search_dn_attribute` does not exist
When running a `search+bind` authentication, we need to get the user DN
from one of the attributes of the LDAP entry of that user. The
`ldap.search_dn_attribute` configuration options specifies which entry
attribute contains the DN.

When the `ldap.search_dn_attribute` is set with an attribute that does
not exist in the LDAP entry, then we need to fail the authentication and
log that information.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>