Release workflow
This commit introduces the instructions for making releases of this
module.

It improves the GitHub action workflow to create the GitHub release
automatically based on the changelog present in the repository.

The instructions for the release manager to follow are present in
`HOWTORELEASE.md` file.

This commit also changes the version of the module to `1.0.0-dev` to
match the new release process rules.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>