Replaced `LDAP.STATUS` command by INFO section
In this commit we remove the `LDAP.STATUS` in favor of adding a new
section to the `INFO` command to show the same information about the
LDAP servers' connection health.

The new section is named `ldap_status` and it's content is a set of
dictionaries, one per LDAP server, and in each dictionary, we have the
`url`, `status`, `ping_time`, and `error` fields.

Example of the output:

```
 > INFO ldap
 # ldap_status
 ldap_server_0:url=ldap://ldap,status=unhealthy,error=LDAP connection failure: op send error: channel closed
 ldap_server_1:url=ldap://ldap-2,status=healthy,ping_time(ms)=1.645
```

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>