Run integration tests against several Valkey versions with ASAN enabled
This commit makes changes to the integration testing infrastructure to
enable running integration tests against the different supported valkey
versions, namely 7.2, 8.0, and 8.1.

We also enable ASAN compilation of Valkey server, to detect memory
errors.

When we enabled ASAN, we immediately found a memory error in the logging
system, and we also fix the problem in this commit.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>