Make valkey an RPM weak dependency of valkey-ldap module
This commit changes the type dependency between the valkey-ldap module
and the valkey server in the RPM spec file to be a weak dependency.

We're making valkey server to be a weak dependecy because sometimes
users might install valkey server without using the OS package manager,
and we still want the user to be allowed to install valkey-ldap module
using the OS package manager even if valkey is not detected as
installed.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>