Fix shutdown of tokio runtime
This commit fixes the shutdown of the tokio runtime when the module is
unloaded. Before this commit, the tokio runtime was not being shut down,
and pending async tasks were still running, which would cause illegal
memory accesses when logging messages because the logging system was
using a no longer valid Valkey context structure.

This should fix the flaky test errors.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>