[backport] Fix CONFIG REWRITE from always re-writing the `ldap.search_bind_passwd` config
The current behavior of `CONFIG REWRITE` is that it will always write
an obfuscated value for the `ldap.search_bind_passwd` config in valkey.conf
even if `ldap.search_bind_passwd` has not been updated.

This behavior is due to a strategy of making sure that the value of
`ldap.search_bind_passwd`, loaded from valkey.conf when the server
starts, does not show up in `CONFIG GET` operations. To achieve this,
the module implementation overrides the update procedure of the
`ldap.search_bind_passwd`, and stores the password in a temporary
variable and writes `******` in the `ldap.search_bind_passwd` config
variable.

The side-effect of this implementation is that `CONFIG REWRITE` thinks
that the config has changed and writes the current obfuscated value
`******`.

This commit fixes the problem by abandoning the strategy of obfuscating
the config value, and adds teh `HIDDEN` flag to the config variable
declaration to prevent `CONFIG GET ldap.*` command from showing the
password value.

Fixes: #52

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>
(cherry picked from commit bc46f7258aa1f9b1338e6971a5750111eec34e00)